Started working on the summarized plot.
This view of the data is horizontal instead of vertical and summarizes data instead of showing every data point for every sample. This commit includes an update to the global layout variables, the toolbar, interactivity and the plotting functions. So far, only the coordinates are drawn when creating the summarized plot.