Moved the genomic annotation to the bottom margin.
Since we will have to plot numeric values (e.g. DNA methylation values) on the x axis instead of genomic coordinates, it is easier to push the genomic annotation to the bottom margin. This way, we can just plot the values without having to take into account the space taken by the genomic annotation.