Changed the visibility animation of the toolbar.
Originally, the toolbar had an opacity of 0 before a plot was made, so that it was not visible, but still filled the necessary space in the DOM. The problem was that buttons could still be clicked, even when the toolbar was not visible, leading to all sorts of errors. Now, an extra div has been added to cover the toolbar. This cover is then simply faded out when we need the toolbar and we no longer have to mess with the toolbar opacity.