Fixed error when opening downloaded .svg file in Illustrator.
D3.js automatically adds a font-family attribute to an axis with the value "sans-serif". Illustrator does not recognize this value and therefore cannot open the .svg file (see here: http://x3ro.de/solving-svg-import-issues-in-adobe-illustrator-cs6-and-cc/). By specifying a font-family that Illustrator can recognize (arial), the file can be opened in Illustrator.