Fixed closed line path bug.
The d3.line function automatically adds a "Z" to the end of the path data, which closes the path by drawing a line from the last point back to the first. By specifying the fill attribute of the path to be 0, the path does not appear to be closed in the browser. To make sure the path is also open when looking at the SVG in Adobe Illustrator, we have to remove the trailing "Z" from the path data before drawing it.