Using the filtered data instead of all data.
Instead of having a mixed usage of the full dataset (the cancerTypeData object) and the filtered dataset (the cancerTypeDataFiltered object), all processing, calculating and plotting is now performed on the filtered dataset. The only times the full dataset is now is when the plot is first drawn or when the user resets the plot. In these cases the full dataset is simply copied to the variable that holds the filtered dataset.