Fixed large number errors.
Some of the statistical functions, such as beta() and fDistribution(), would generate very large numbers in their calculations. Whenever these large numbers exceeded the largest possible number in JavaScript (1.7976931348623157e+308), the calculations would result in Infinity and the eventual p value would be NaN. We have added the math.js library to MEXPRESS to be able to deal with extremely large numbers and avoid the aforementioned problem.