Added check for availability of Object.values method.
Without this function, MEXPRESS simply does not work. Older versions of recommended browsers such as Firefox or Chrome also do not contain this function, which is why I added this extra check. This update also includes some layout changes to the error message shown when the browser is not supported.