Added ovarian cancer.
Ovarian cancer is a special case in that almost all DNA methylation data was measured on the 27k platform instead of the 450k one. So I added a test to the data download script that checks whether the requested cancer type is ovarian cancer or not. If it is, the 27k data is returned, if not, the 450k data. I also removed all references to the 450k platform in the variable names.