Fixed p value bug
Whenever the distributions of the input data for the anova function were too similar, the estimated p value would go over 1. This is of course statistically impossible and purely the result of the numerical approximation. To solve this problem, we now simply check whether the p value is higher than 1 and setting it to 1 if it is.